fix: quality hardening — atomic writes, path traversal, panics, PRAGMAs, clippy, SPDX
- Make ChunkStore::put() atomic with write-to-temp-then-rename
- Reject path traversal (../) and absolute paths on tar ingest
- Replace bare .unwrap() on SystemTime with descriptive .expect()
- Replace panicking .expect() in get_file_chunks with proper error
- Add SQLite PRAGMAs (WAL, synchronous, foreign_keys, busy_timeout)
- Fix all clippy warnings (unnecessary casts, io_other_error, manual_strip)
- Replace magic 20MB satellite read limit with named MAX_CHUNK_RESPONSE
- Add SPDX-License-Identifier: AGPL-3.0-only to all source files
- Add license, description, repository to all Cargo.toml files

39 tests passing (4 new: atomic write, path traversal x2, WAL mode).

+++ crates/rk-core/src/chunker.rs
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
 use crate::chunk_store::ChunkStore;
 use crate::Result;
 
@@ -27,7 +29,7 @@ pub fn chunk_data(data: &[u8], min_size: u32, avg_size: u32, max_size: u32) -> V
     let chunker = FastCDC::new(data, min_size, avg_size, max_size);
     chunker
         .map(|entry| {
-            let slice = &data[entry.offset as usize..entry.offset as usize + entry.length];
+            let slice = &data[entry.offset..entry.offset + entry.length];
             ChunkMeta {
                 hash: blake3::hash(slice),
                 offset: entry.offset as u64,
